Switch to using zplug instead of having zprezto repo as submodule
The idea is to be able to quickly try other plugins, clean up my repo a
bit, and get a more up to date terminal. However, there seem to be quite
some speed issues. This is why there's also some antigen and zgen
prototyping in the zshrc file. For now let's see how this works.